requirements.txt
fastapi
uvicorn
openpyxl
pandas
pytesseract
Pillow
groq
camelot-py[cv]
opencv-python
ghostscript
python-multipart
